Make major revisions on drawing functionalities
New changes were made involving the drawing behavior where holding left click would
allow the user to draw compare to before where users are able to draw just by
merely hovering over the canvas which is not good if the user wants to draw at the
middle of the canvas. As a result of this, new functions were added to track the
drawing state when ever the left mouse button is being held.